refactor: extraced game structure
