+++ src/game/constants.ts
@@ -1,0 +1,4 @@
+export const PLAYER_START_POS = 50;
+export const PLAYER_STEP = 5;
+export const PLAYER_MIN = 0;
+export const PLAYER_MAX = 100;
